Updated gitignore, index, README, FlowArray, MainTest
.gitignore was updated to include more eclipse specific directories and files.
Added update 10 to index.
README was updated to include a table of omitted files + folders and to also include a clause at the bottom which says not to edit the .gitignore without changing the table.
Update 10 to FlowArray fixes a problem with initialising typeless FlowArrays.
MainTest is changed to reflect this.